Cleaned up future public API
Added abstraction layer between "MockitoFramework" type and the logic that creates invocations.
This will be useful in the future if there are more creation methods.
It also makes more sense from the standpoint of abstractions - "MockitoFramework" is top level API for framework integrations and should contain only high level elements.